Improve method bar and violin
