+++ cellrank/tools/estimators/_property.py
@@ -19,12 +19,11 @@ from anndata import AnnData
 import cellrank.logging as logg
 from cellrank.tools import Lineage
 from cellrank.utils._docs import d
-from cellrank.tools._utils import _make_cat, partition, _complex_warning
+from cellrank.tools._utils import RandomKeys, _make_cat, partition, _complex_warning
 from cellrank.tools.kernels import PrecomputedKernel
 from cellrank.tools._constants import Direction, DirPrefix, DirectionPlot
 from cellrank.tools.estimators._utils import (
     Metadata,
-    RandomKeys,
     _create_property,
     singledispatchmethod,
     _delegate_method_dispatch,
