+++ tox.ini
@@ -9,7 +9,8 @@ filename = *.py
 # W503 line break before binary operator
 # D105 Missing docstring in magic method
 # E203 whitespace before ':'
-ignore = D202,D107,B008,W503,D105,E203
+# F405 ... may be undefined, or defined from star imports: ...
+ignore = D202,D107,B008,W503,D105,E203,F405
 exclude =
     .git,
     __pycache__,
